+++ Swift-SourcesHeaderChecker/Swift-SourcesHeaderChecker/Checker/HeaderVerifier.swift
@@ -21,14 +21,13 @@
      SOFTWARE.
  */
 
-
 import Foundation
 
 /// Structure which has the aim of reading files and check if they contain in the top the suitable
 /// leagal mentions in the header.
 ///
 /// - Author: Pierre-Yves Lapersonne
-/// - Version: 1.0.0
+/// - Version: 2.0.0
 /// - Since: 01/07/2019
 ///
 struct HeaderVerifier {
@@ -44,22 +43,38 @@ struct HeaderVerifier {
     /// - Parameters:
     ///     - at: The array of files' paths which must contain the header
     ///     - for: The content of the header to get in the top of the files
+    ///     - ignoring: The number of lines to ignore starting from the top of the processed file
+    ///     - excluding: The path to file containing files' paths to ignore for the program
     /// - Returns:
     ///     - A boolean value indicating if the header is available in all the files (true)
     /// or not (false if at least 1 file)
     ///
-    func look(at files: [String], for mention: String) -> Bool {
-        output.verbose("Will process \(files.count) files")
+    func look(at files: [String], for mention: String, ignoring lines: Int,
+              excluding protected: String  = "") -> Bool {
+        
+        output.verbose("Will process \(files.count) file(s)")
+        
+        let protectedFiles = protected.lines.filter({ !$0.clear().isEmpty })
+        if protected.linesCount > 0 {
+            output.verbose("⚠️  \(protectedFiles.count) file(s) will be ignored")
+        }
+        
         var allFilesAreSuitable = true
         for file in files {
-            output.verbose("Processing file '\(file)")
-            let isFileSuitable = look(at: file, for: mention)
-            if !isFileSuitable {
-                output.write("❌ It seems the file at \(file) does not have a suitable header")
+            if protectedFiles.contains(file) {
+                output.verbose("⚠️  Ignoring file '\(file)")
+            } else {
+                output.verbose("Processing file '\(file)")
+                let isFileSuitable = look(at: file, for: mention, ignoring: lines)
+                if !isFileSuitable {
+                    output.write("❌ It seems the file at \(file) does not have a suitable header")
+                }
+                allFilesAreSuitable = allFilesAreSuitable && isFileSuitable
             }
-            allFilesAreSuitable = allFilesAreSuitable && isFileSuitable
         }
+        
         return allFilesAreSuitable
+        
     }
     
     /// Reads the content of the file at this path and look for the mention in its head.
@@ -72,16 +87,18 @@ struct HeaderVerifier {
     /// Will consider as suitable a file which has its header matching the following conditions:
     /// 1 - no hashbang in first line (because of 3)
     /// 2 - no import in first line (because of 3)
-    /// 3 - starts with /** or /*
-    /// 4 - has */ ending the header
+    /// 3 - starts with /** or /* or //
+    /// 4 - has */ ending the header or //
     ///
     /// - Parameters:
     ///     - at: The path to the source file which should contain the expected header
     ///     - for: The header content to find
+    ///     - ignoring: The number of lines to ignore starting from the top of the processed file
     /// - Returns:
-    ///     - A boolean value indicating if the header is available (true) or not (false)
+    ///     - A boolean value indicating if the header is available (true) or not (false), or if
+    /// the file to process is not suitable (false)
     ///
-    func look(at path: String, for mention: String) -> Bool {
+    private func look(at path: String, for mention: String, ignoring lines: Int) -> Bool {
         
         // Read the current file
         guard let currentFileContent = try?
@@ -90,30 +107,57 @@ struct HeaderVerifier {
             return false
         }
 
-        // Check if the line starts by /** or //
-        if !currentFileContent.starts(with: "/**") && !currentFileContent.starts(with: "/*") {
-            output.write("⚠️  It seems the Swift file at (\(path)) does not start with '/**' or '/*'. Will reject it.")
+        // Check if enough lines are ignored compared to the under-process file size
+        if lines >= currentFileContent.linesCount {
+            output.write("⚠️  The number of lines to ignore is greater than the number of lines of the file to process",
+                         to: .error)
             return false
         }
         
+        // Check if the line starts by /**, /* or //
+        if !currentFileContent.matchStartCommentLine() {
+            output.write("⚠️  It seems the Swift file at (\(path)) does not start with '/**' or '/*' or '//'. Will reject it.")
+            return false
+        }
+
+        // Reduce the file using ignored lines
+        // FIXME Dirty!
+        let topRecudedFileContent = currentFileContent.lines.suffix(from: lines).joined(separator: "\n")
+        
+        // Keep just enough lines
+        let bottomReducedFileContent = topRecudedFileContent.linesUntil(k: mention.linesCount + 2) // Keep ending line
+        
         // Compare line by line
-        // FIXME Can we optimize this part with low-cost methods or standard API?
-        let splittedFileContent = currentFileContent.linesUntil(k: mention.linesCount + 2) // Keep */
         let splittedMention = mention.lines
         for i in 0..<mention.linesCount {
-            let cleanedMentionLine = splittedMention[i].clear()
-            let cleanedFileContentLine = splittedFileContent[i+1].clear()
+            
+            let currentFileLine = bottomReducedFileContent[i]
+            let currentMentionLine = splittedMention[i]
+            
+            var cleanedFileContentLine = "", cleanedMentionLine = ""
+            
+            // Deal with case of final line with only //
+            if currentFileLine.count > 0 && currentFileLine != currentMentionLine {
+                cleanedFileContentLine = currentFileLine.clear()
+                cleanedMentionLine = currentMentionLine.clear()
+            } else {
+                cleanedFileContentLine = currentFileLine
+                cleanedMentionLine = currentMentionLine
+            }
+            
             if cleanedMentionLine != cleanedFileContentLine {
-                output.write("❌ The line \(i+1) did not match between the file \(path) and the header. Will reject it.")
-                output.write("\t - Here the line \(i+1) of the header: '\(splittedMention[i])'")
-                output.write("\t - Here the line \(i+1) of the current file: '\(splittedFileContent[i+1])'")
+                output.write("❌ The line \(i) did not match between the file \(path) and the header. Will reject it.")
+                output.write("\t - Here the line \(i) of the header: '\(currentMentionLine)'")
+                output.write("\t - Here the line \(i) of the current file: '\(currentFileLine)'")
                 return false
             }
+            
         }
         
-        if splittedFileContent.last?.clear() != "*/" {
-            output.write("⚠️  It seems this Swift file (\(path)) has its header closed by another symbol than */")
-        }
+        // TODO: Deal with end by //
+//        if !(bottomReducedFileContent.last?.matchEndCommentLine())! {
+//            output.write("⚠️  It seems this Swift file (\(path)) has its header closed by another symbol than */ or //")
+//        }
         
         return true